Don’t record temporary window selection changes
There are places with code similar to the following:

      (save-selected-window
        (select-window window)
        …)

For the purposes of this commit, this is roughly equivalent to:

      (let ((old-window (selected-window)))
        (select-window window)
        …
        (select-window old-window 'norecord))

That is, when ‘save-selected-window’ restores the state, it passes
‘norecord’ argument to ‘select-window’ function.  As a result, when
selected window is initially changed ‘buffer-list-update-hook’ is
called however it is not invoked when selected window is restored.
This confuses any code using that hook to track selected window.

To address this issue, change calls such as the above to use
‘norecord’ argument.  This way, ‘buffer-list-update-hook’ is called
neither when the new window is temporarily selected nor when old
window is restored.

For reference, this is indeed what documentation of ‘select-window’
says should be done:

> Run ‘buffer-list-update-hook’ unless NORECORD is non-nil.  Note that
> applications […] often select a window temporarily […]  to simplify
> coding.  As a rule, such selections should not be recorded and
> therefore will not pollute ‘buffer-list-update-hook’.  Selections
> that "really count" are those causing a visible change in the next
> redisplay of WINDOW’s frame and should always be recorded.

Fixes: https://github.com/mina86/auto-dim-other-buffers.el/issues/26