`pdf-info-renderpage`: Use the width of the page as passed in.
`pdf-view-image-size` passes in the width of the current page to
`pdf-info-renderpage` during an isearch. Use this as-is, instead of
multiplying it with the `pdf-util-frame-scale-factor`.

This is a potential fix for the bug raised by PR #13.

I'm not sure if this was a bug earlier as well, but was masked since
the code-path did not execute for `png` images.

I will treat this as a temporary change until I can better understand
the original reasons behind putting it in. Since I don't have access
to a Mac running `emacs-mac` , I have tested it by hard-coding
`pdf-util-frame-scale-factor` to 2 on my local setup, and everything
seems to be working correctly.