Update all the docker template files
This work is part of the effort to fix the `test-all` Make target.
When I complete all the work, I will update the relevant Circle CI
scripts to run these tests as well.

Note also that I personally use podman instead of docker, so I'm
replacing docker with podman in the scripts. Someday, I might find a
clever way to let both co-exist