Add commands to create text annotations in the margins
This adds the commands 'pdf-keynav-annot-add-text-left/right' and a variable
'pdf-keynav-annot-text-margin' to control where the text annotations are placed
in the margins. A function 'pdf-keynav-annot-delete-closest-text' is also added
and command 'pdf-keynav-annot-delete' is modified to incorporate it. Furthermore,
an error in 'pdf-keynav-annot-activate' has been fixed, plus custom types have
been adjusted and various formatting fixed.