Sanitize buffer name for use as temp file prefix
Prefixes for temporary files are derived from either the file name
associated with the pdf-view buffer, or the buffer name if it is not
visiting a file.  If the buffer name contains '/' characters, then it is
not a valid file name.

Clean up non-alphanumeric characters in the buffer name before using
it as a temp file prefix.