Use `l` and `r` for backward / forward in history.
`l` and `r` are the conventional bindings for history navigation.

The previous key-bindings `B` and `N` are kept as-is for people who
were used to it.

This is a breaking change because `r` was previously bound to
`revert-buffer`. However, `g` is also bound to `revert-buffer` and is
the conventional binding for `revert` so this should be okay.

Fixes: #43