Guard against dead buffers in pdf-tools-install-noverify
I am not sure how to 'consistently' reproduce the issue that this
commit fixes. But I was experiencing the issue as reported at
https://github.com/syl20bnr/spacemacs/issues/15106, where for some
reason `pdf-view-mode` did not open in evil `evilified-state`.

After (a long time of) hunting down the cause, I found that
`pdf-tools` (silently) errors when the `buffer-list` contains killed
buffers. The error occurs because `pdf-tools-install-noverify` calls
`with-current-buffer` with the killed buffer. This commit simply adds
the `buffer-live-p` check before the `with-current-buffer` form gets
called.

Closes: #93
Potentially Fixes: syl20bnr/spacemacs#15106
Author: @dalanicolai