Add workaround for compilation using C++17 standard library
As explained in #102 by @Pi-Cla and @uliw, `pdf-tools` relies on
private headers provided by `poppler`. In order to compile these
correctly, we need the standard library provided by C++17.

`pdf-tools` is built and tested against C++11. This workaround
provides the correct CXXFLAGS to compile the library and fixes the
immediate broken compilation problem.

There should be two "correct" fixes to these:
1. Remove dependence on private `poppler` headers. This is being
tracked in #103 and #105
2. Upgrade `pdf-tools` to build and run against new C++ versions. This
is being tracked in #109.

Closes: #70, #76, #102, #108, politza/pdf-tools#706.