Revert "Use mkstemp instead of tempnam"
This reverts commit d63a1e7d87f9b0a19209f2eeb170bcf64612aa2f.

In d63a1e7, @JunyuanChen fixed a long standing compilation warning by
replacing `tempnam` with `mkstemp` in `epdfinfo`. However, the
`mkstemp` implementation does not work correctly on MS
Windows (probably because the path template is "wrong" for Windows).

I am reverting the commit and opening a new issue #110 to track the
correct implementation of `mkstemp` (or equivalent) on MS Windows.

It would also be great to add a test against Windows CI to open a PDF
and check that the operation completes successfully.

Closes: #101