Use mkstemp instead of tempnam
This reverts commit fedd930a09a497c03df3ce5204ccbd80da724662. It
should be possible to use the filename as generated by `mkstemp`, the
problem was that we were using the wrong path-separator.

So the diff from fedd930a09a497c03df3ce5204ccbd80da724662 is as
follows:

@@ -338,7 +338,11 @@ strchomp (char *str)

   return str;
 }
-
+#if HAVE_W32
+#define MKTEMP_SEPARATOR "\\"
+#else
+#define MKTEMP_SEPARATOR "/"
+#endif
 /**
  * Create a new, temporary file and returns its name.
  *
@@ -347,7 +351,7 @@ strchomp (char *str)
 static char*
 mktempfile()
 {
-  char template[] = P_tmpdir "epdfinfo_XXXXXX";
+  char template[] = P_tmpdir MKTEMP_SEPARATOR "epdfinfo_XXXXXX";

I have tested the change on Mac OSX and Linux. It is tested on Windows
10 by @ShuguangSun.

Closes: #110