Disable pixel-scroll-precision-mode locally if enabled
This change disables the emacs 29 `pixel-scroll-precision-mode` (if
enabled) which interferes with the `pdf-tools` scrolling mechanism. I
think the `pixel-scroll-precision-mode` does not make sense in the
context of the `pdf-view` mode, which does pixel based scrolling
anyways.

After this change, scrolling in a `pdf-view` buffer should work again
with `pixel-scroll-precision-mode` enabled.

Closes: #56
Closes: #124