Fix: Handle nil return value from `image-scroll` functions
*Note*: This is a temporary fix.

What happened?

When the `arg` to `image-scroll-up` is 0, the function disregards the
argument (scrolling up 0 lines means not changing anything) and
returns a `nil` value. This causes the equality check in this commit
to fail (the check is to see if we have hit the top or bottom of the
page, since in this case "moving the image" will do nothing and we
will have to move the page).

The current fix adds an explicit null check to ensure that we don't
compare scroll positions.

Unknowns: under what conditions is the argument to
`pdf-view-scroll-up-or-next-page` 0?

Things to do: Review the `= window-vscroll image-scroll` pattern all
over the code-base and refactor it to deal with `arg = 0` problem.

Fixes: #131