Address some byte-compiler warnings
* lisp/pdf-annot.el (pdf-annot-run-modified-hooks): Use cl-case
instead of cl-ecase, as the latter does not support t/otherwise
clauses.  Remove nil clause which never matches.
See https://bugs.gnu.org/51368.
(pdf-annot-edit-contents-finalize):
* lisp/pdf-sync.el (pdf-sync-minor-mode): Mark up symbols
`like-this', to avoid having the apostrophe interpreted as a closing
single quote.  Remove redundant backslashes.
(pdf-sync-synctex-file-name): Use line-beginning-position in place
of point-at-bol, which is deprecated in Emacs 29.
* lisp/pdf-util.el: Autoload list-colors-duplicates before its use
in pdf-util-color-completions, as facemenu is not preloaded since
Emacs 28.
(pdf-util-hexcolor): Use ash in place of lsh, which is deprecated in
Emacs 29.
(pdf-util-convert): Fix docstring markup and indentation.
* lisp/pdf-virtual.el (pdf-virtual-define-adapter): Refill generated
docstring to avoid exceeding 80 column warning threshold.