Speed up server testing by keeping the containers around
This allows re-using containers, saving a tremendous amount of time.

The drawback here is that any environment change (on the docker /
podman containers side) is not detected. The way to deal with this
will be to create a process around major releases where I delete all
the containers and re-test from scratch.

Relates to: #130