feat: support pdf-buffer bookmarks in workspace managers
Don't assume that there is a window displaying the buffer when
`bookmark-jump` is called. Instead of `bookmark-after-jump-hook` use
`window-buffer-change-functions` so that the buffer is actually
displayed when the code assuming a window is run.

If `window-buffer-change-functions` isn't bound fallback to
`bookmark-after-jump-hook`. The window argument for the inserted hook
is optional to accommodate both cases.

Use `window-buffer-change-functions` only if buffer has not window.
Save the same bookmark if it wasn't restored

Closes: #188
Author: @aikrahguzar