Allow creation of markup annotations without a text selection
This allows calling `pdf-info-addannot` from Lisp with a nil
`markup-edges` argument, which was previously disallowed.

In this case, the annotation covers exactly the selected region of the
page, without attempting to match it to a certain text region.

This means that you can now highlight an exact region! Just select the
region by holding down the Meta key and dragging the mouse. Then
invoke `pdf-annot-add-highlight-markup-annotation` (with `C-c C-a h`)
and voila!

Closes: #191