keep cursor on the same line when reopening list of annotations
Simply check whether the buffer of the list of annotations already
exists.  In that case, avoid calling `tabulated-list-print` because
the call to that function makes the cursor move to the beginning of
the buffer.  Tested for the case in which the buffer of the list of
annotations already exists; a new note is created; that buffer is
reopened with the cursor on the same position as previous left; and
the newly inserted note appears as expected.  Not sure if further
tests should be done or the implication of avoiding calling
`tabulated-list-print` every time.