ensure it works when opening list, closing PDF and reopening PDF
Previous commit has a bug: it was not properly opening the list of
annotation when opening PDF; opening the list of annotations;
keeping the buffer containing the list of annotations; closing the
PDF; reopening the same PDF; and trying to reopen the list of
annotations.

This commit fix that bug, being able to open the list of
annotations after the PDF is closed and reopened.