Bump pdf-tools version to v1.1.0
This is necessary because the extensive changes in `epdfinfo` do not
get picked up by third party package management systems unless they
enter Melpa Stable.

This mismatch caused a number of issues, which I expect this bump will
now fix. Thanks to @cyrilarnould and @barslmn for pointing this out.

I am also using this bump to close long-open install issues. The idea
here is that these issues are being lost. I would rather everyone
upgrade to the latest versions of `pdf-tools` and `epdfinfo` and then
re-report the issues if they still exist.

Please note: as of this release, users are not expected to manually
set anything on their machines during installation. (for example:
`$PKG_CONFIG_PATH`). The installation should *just work*. If it does
not, please open a ticket so that we can track it.

Fixes: #213
Closes: #177
Closes: #216