Update NEWS with recent contributions
The goal is to try and keep these updates as close to when I merge the
changes as possible, since they become unmanageable if you wait until
the day of the release.

I want to acknowledge all the amazing people people who are
contributing to `pdf-tools` and making it the powerhouse it is.