[v2] Auto enable pdf-view-dark-minor-mode if the theme is dark
The first attempt at this workaround was
c451c0e3db5b367d191bddf428e38fd8a21c0102 , which did not work
because (my guess) `pdf-tools-enabled-modes` would be evaluated before
the theme was loaded. This is attempt 2, which sets this value in the
`pdf-view-mode` function itself.

This is meant as a workaround for #166 (which see).

Best guess about what is happening and why this workaround works comes
from @d8660091 (lightly edited):

> I'm using M2 and I can reproduce the issue with just the following
> snippets I extracted from pdf-isearch-current-colors, which was
> invoked during a pdf isearch:

    (add-hook 'pdf-tools-enabled-hook
               (lambda ()
                 (pdf-util-face-colors 'pdf-isearch-match nil)
                 (pdf-util-face-colors 'pdf-isearch-lazy nil)))

> And run `pdf-tools-install`

> I tried adding `(sit-for 1)` before `(delete-frame f)`, the crash
> was gone.

> But I am still not sure what's really happening.

> I think `pdf-view-dark-minor-mode` works because new frame is not
> created and `delete-frame` doesn't happen.

If anyone has an idea about what is happening here and what the
correct way to fix it is, please get in touch via the linked issue!