Pass window argument + fix arithmetic condition
There was a bug where the pages won't display because of not
not passing the window argument.

There was was also the problem that page would fail to render if
the remaining portion was exactly equal to step-size.

Lastly do work in pre-display whenever possible. So in
`image-mode-redisplay` we just mark window for redisplay
by calling `force-window-update`