Accommodate image-mode commands using an advice
`pdf-view-mode-map` uses `image-mode-map` as its parent map.
These commands are used for various things including horizontal
scrolling. They look for the image at `point-min` which is not
the case for pdf-view-roll-minor-mode. So we add a before-until
advice that checks for this minor mode and returns the correct
image if it is active.

This approach might need to be revisited. An alternative is to
use narrow the buffer but it can have many unintended consequences.
I think advice is better because it is very targeted.