Refactor bookmark handling
Self removing hook is no more. Instead a buffer local hook
is now used to set vscroll after jumping to a bookmark.

The same function is run the very first time a buffer is
displayed. This takes care of the scenario where a new buffer
is not immediately displayed upon jumping a bookmark.