Aggressively flush images
This ensures that we don't use lot of memory in
scenarios such as the following:
1) Scrolling through lot of pages
2) Frequent reverting e.g. when writing a TeX documents
3) Frequent change of buffers

I have seen image cache grow as large as 5-6 gb in such
circumstances.

For me this change doesn't seems to impact performance
while it reduces memory pressure. However, this might not
be the universal experience so this change might be
revisited in the future.