Don't try to find origin of buffers not displayed in a live window
This is a bit too involved since finding the dimension of the
displayed image is a bit complicated in this case.

I would like this to be possible since this means that buffers
that are being displayed in a non-current tab can be bookmarked
but that requires a bit more thought.

The current approach resulted in errors for non-current tabs.